NXT-511 Split layout into server and client components
